Refactor ECS cluster configuration and variables
This commit refactors the main.tf file to update the subnet configuration for the ECS cluster instances. Instead of using a specific subnet, the instances will now use the subnets specified in the var.subnets variable. Additionally, the ami variable is formatted for consistency.

In the variables.tf file, a new variable "cluster_size" is added with a default value of 3.

These changes improve the flexibility and maintainability of the ECS cluster configuration.